Refactor API calls into subfiles
